Progress commit: работает

+++ test-application/src/main/java/ru/nsu/service/BatchCreator.java
@@ -6,10 +6,10 @@ import ru.nsu.cloud.model.mapreduce.MapFunction;
 import java.util.Collection;
 import java.util.List;
 
-public class BatchCreator implements MapFunction<List<Long>> {
+public class BatchCreator implements MapFunction<List<Integer>> {
 
     @Override
-    public Collection<List<Long>> apply(List<Long> longs) {
+    public Collection<List<Integer>> apply(List<Integer> longs) {
         return Lists.partition(longs, 10);
     }
